change gentx to include adversarial

+++ src/gentx.py
@@ -21,13 +21,13 @@ def _create_random_transactions(message_client):
         batch_txns = messages_pb2.BatchTransactions()
         entries = 100
 
-        for i in range(entries/2):
+        for i in range(int(entries/2)):
             recipient_addr = random.choice(
                 message_client.state.get_all_addresses())
             msg_obj = message_client.generate_txn_object(recipient_addr, 100)
             batch_txns.transactions.append(msg_obj)
 
-        for i in range(entries/2):
+        for i in range(int(entries/2)):
             conflict_msg_obj = message_client.generate_conflicting_txn(
                 msg_obj, attacker.address, 100)
             batch_txns.transactions.append(conflict_msg_obj)
